# magma-fusion-impl/src/main/java/fr/insee/rmes/magmafusion/api/DatasetsEndpoints.java
package fr.insee.rmes.magmafusion.api;

import fr.insee.rmes.magmafusion.api.requestprocessor.RequestProcessor;
import fr.insee.rmes.magmafusion.model.Dataset;
import fr.insee.rmes.magmafusion.model.Distribution;
import fr.insee.rmes.magmafusion.queries.parameters.DatasetsRequestParametizer;
import fr.insee.rmes.magmafusion.services.DatasetsService;
import fr.insee.rmes.magmafusion.utils.DatasetByIdDTO;
import fr.insee.rmes.magmafusion.utils.DatasetDTO;
import fr.insee.rmes.magmafusion.utils.DistributionDTO;
import org.springframework.http.ResponseEntity;
import org.springframework.lang.Nullable;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
public class DatasetsEndpoints implements JeuxDeDonneesApi {

    private final RequestProcessor requestProcessor;
    private final DatasetsService datasetsService;

    public DatasetsEndpoints(RequestProcessor requestProcessor, DatasetsService datasetsService) {
        this.requestProcessor = requestProcessor;
        this.datasetsService = datasetsService;
    }

    @Override
    public ResponseEntity<List<Dataset>> getListDatasets(@Nullable String dateMiseAJour) {
        List<DatasetDTO> dtos = requestProcessor.queryToFindAllDatasets()
                .with(new DatasetsRequestParametizer(dateMiseAJour))
                .executeQuery()
                .listResult(DatasetDTO.class)
                .result();

        List<Dataset> dataSets = datasetsService.convertDatasetDTOsToDataSets(dtos);
        return ResponseEntity.ok(dataSets);
    }

    @Override
    public ResponseEntity<Dataset> getDataSetById(String id) {
        DatasetByIdDTO dto = requestProcessor.queryToFindDatasetById()
                .with(new DatasetsRequestParametizer(id, null))
                .executeQuery()
                .singleResult(DatasetByIdDTO.class)
                .result();
        if (dto == null) {
            return ResponseEntity.notFound().build();
        }
        return ResponseEntity.ok(datasetsService.convertDatasetByIdDTOToDataSet(dto));
    }

    @Override
    public ResponseEntity<List<Distribution>> getDataSetDistributionsById(String id) {
        List<DistributionDTO> dtos = requestProcessor.queryToFindDistributionsByDatasetId()
                .with(new DatasetsRequestParametizer(id, null))
                .executeQuery()
                .listResult(DistributionDTO.class)
                .result();
        return ResponseEntity.ok(datasetsService.convertDistributionDTOsToDistributions(dtos));
    }
}