# magma-fusion-oas/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: MagmaFusion API
  version: ${version}
  description: API sur les métadonnées de l'Insee
  contact:
    name: dr59-sndil
    url: https://www.insee.fr/fr/accueil
    email: dr59-sinl-equipe-maintenance-rmes@insee.fr
...
security:
  - bearerAuth: []

### TAGS SPECIFICATION ###
...
paths:
...
  /datasets/list:
    get:
      tags:
        - Jeux de donnees
      summary: Get list of datasets
      operationId: getListDatasets
      parameters:
        - name: dateMiseAJour
          in: query
          description: 'Date of last update. Example: 2023-01-31'
          required: false
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/DatasetsListe200'
        '400':
          $ref: '#/components/responses/IncorrectRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '500':
          $ref: '#/components/responses/ServerError'
  /dataset/{id}:
    get:
      tags:
        - Jeux de donnees
      summary: Get one dataset
      operationId: getDataSetById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Dataset200'
        '400':
          $ref: '#/components/responses/IncorrectRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '500':
          $ref: '#/components/responses/ServerError'
  /dataset/{id}/distributions:
    get:
      tags:
        - Jeux de donnees
      summary: Get dataset distributions
      operationId: getDataSetDistributionsById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/DistributionsListe200'
        '400':
          $ref: '#/components/responses/IncorrectRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '500':
          $ref: '#/components/responses/ServerError'

### COMPONENTS SPECIFICATION ###
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  ### PARAMETERS SPECIFICATION ###
  responses:
    IncorrectRequest:
      description: 'La syntaxe de la requête est incorrecte'
    NotFound:
      description: 'Ressource non trouvée'
    NotAcceptable:
      description: 'L''en-tête HTTP ''Accept'' contient une valeur non acceptée'
    ServerError:
      description: 'Erreur interne du serveur'
...
    Dataset200:
      description: 'Opération réussie'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Dataset'
          examples:
            défaut:
              $ref: '#/components/examples/Dataset'
        application/xml:
          schema:
            $ref: '#/components/schemas/Dataset'
          examples:
            défaut:
              $ref: '#/components/examples/Dataset'
    DistributionsListe200:
      description: 'Opération réussie'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ListeDistributions'
          examples:
            défaut:
              $ref: '#/components/examples/ListeDistributions'
        application/xml:
          schema:
            $ref: '#/components/schemas/ListeDistributions'
          examples:
            défaut:
              $ref: '#/components/examples/ListeDistributions'
    DatasetsListe200:
      description: 'Opération réussie pour les datasets'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ListeDatasets'
          examples:
            défaut:
              $ref: '#/components/examples/ListeDatasets'
        application/xml:
          schema:
            $ref: '#/components/schemas/ListeDatasets'
          examples:
            défaut:
              $ref: '#/components/examples/ListeDatasets'
...
  examples:
...
    Dataset:
      value:
        code: '25baaf1f-61cf-4d50-9124-1477afca2aa4'
        uri: 'http://bauhaus/catalogues/jeuDeDonnees/25baaf1f-61cf-4d50-9124-1477afca2aa4'
        catalogRecordCreated: '2024-12-09T12:00:00.000'
        catalogRecordModified: '2024-12-09T12:00:00.000'
        catalogRecordCreator: 'http://bauhaus/organisations/insee/HIE2004937'
        identifier: 'DD_EEC_SERIES'
        validationState: 'Provisoire, jamais publiée'
        title:
          - contenu: 'Activité, emploi et chômage - séries longues'
            langue: 'fr'
          - contenu: 'Activité, emploi et chômage - séries longues'
            langue: 'en'
        description:
          - contenu: 'Les données sont détaillées par sexe, âge, diplôme et catégorie socio-professionnelle (PCS). Pour l’emploi, les données sont détaillées par type de contrat, composante du sous-emploi, temps de travail (temps complet ou partiel). Plusieurs indicateurs en taux sont disponibles : taux d''activité, taux d''emploi ou taux de chômage.\n\nCe jeu de données regroupe notamment les données annuelles de forte valeur (High Value Dataset ou HVD statistiques) désignées par Eurostat sur les sujets suivants : l''emploi , le chômage et la main d’œuvre potentielle.'
            langue: 'fr'
          - langue: 'en'
        scopeNote:
          - contenu: 'Des corrections sur les données concernant la population totale ainsi que sur l''emploi et la part dans l''emploi selon le plus haut diplôme obtenu, par sexe et âge regroupé, en moyenne annuelle ont été apportées en septembre 2024 afin de rectifier une erreur, portant sur les années 2018, 2019 et 2020 uniquement.'
            langue: 'fr'
          - langue: 'en'
        wasGeneratedBy:
          - contenu: 'http://bauhaus/operations/serie/s1223'
            langue: 'fr'
          - langue: 'en'
    ListeDistributions:
      value:
        - identifier: '6d0d296b-9ec5-42fa-a3d5-62ffe7c74dea'
          created: '2025-02-17T08:13:23.382171614'
          description:
            - contenu: 'fichier csv généré par l''API melodi'
              langue: 'fr'
            - langue: 'en'
          downloadURL:
            -
          modified: '2025-02-17T08:13:23.382171614'
          title:
            - contenu: 'Capacités d''hébergements touristiques'
              langue: 'fr'
            - contenu: 'Tourism capacities'
              langue: 'en'
          uri: 'http://bauhaus/catalogues/distribution/6d0d296b-9ec5-42fa-a3d5-62ffe7c74dea'
    ListeDatasets:
      value:
        - id: '25baaf1f-61cf-4d50-9124-1477afca2aa4'
          uri: 'http://bauhaus/catalogues/jeuDeDonnees/25baaf1f-61cf-4d50-9124-1477afca2aa4'
          catalogRecordCreated: '2024-12-09T12:00:00.000'
          catalogRecordModified: '2024-12-09T12:00:00.000'
          catalogRecordCreator: 'http://bauhaus/organisations/insee/HIE2004937'
          validationState: 'Provisoire, jamais publiée'
          title:
            - contenu: 'Activité, emploi et chômage - séries longues'
              langue: 'fr'
            - contenu: 'Activité, emploi et chômage - séries longues'
              langue: 'en'
        - id: 'a85dab6c-ef0f-4caa-9763-c7f8c840e216'
          uri: 'http://bauhaus/catalogues/jeuDeDonnees/a85dab6c-ef0f-4caa-9763-c7f8c840e216'
          catalogRecordCreated: '2024-12-09'
          catalogRecordModified: '2024-12-09'
          validationState: 'Unpublished'
          title:
            - contenu: 'Activité des branches de l''économie'
              langue: 'fr'
            - contenu: 'Activité des branches de l''économie'
              langue: 'en'




  ### SCHEMAS SPECIFICATION ###
  schemas:
...
    IdLabel:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        label:
          type: array
          items:
            $ref: './langue-contenu.schema.json'


    # ============================================
    # SERIES / OPERATIONS SCHEMAS
    # ============================================
...
    Theme:
      type: object
      properties:
        uri:
          type: string
        label:
          type: array
          items:
            $ref: './langue-contenu.schema.json'
        themeTaxonomy:
          type: string
    LocalisedUrl:
      type: object
      properties:
        lang:
          type: string
        url:
          type: string
    ProcessStep:
      type: object
      properties:
        id:
          type: string
        label:
          $ref: 'langue-contenu.schema.json'
    Temporal:
      type: object
      properties:
        startPeriod:
          type: string
        endPeriod:
          type: string
    Structure:
      type: object
      properties:
        uri:
          type: string
        id:
          type: string
        dsd:
          type: string
    WasDerivedFrom:
      type: object
      properties:
        datasets:
          type: array
          items:
            type: string
        description:
          type: array
          items:
            $ref: './langue-contenu.schema.json'
    ListeDistributions:
      description: 'Liste des distributions'
      type: array
      xml:
        wrapped: true
        name: Definition
      items:
        $ref: '#/components/schemas/Distribution'
    Distribution:
      type: object
      properties:
        identifier:
          type: string
        byteSize:
          type: string
        created:
          type: string
        description:
          type: array
          items:
            $ref: './langue-contenu.schema.json'
        downloadURL:
          type: array
          items:
            type: string
        format:
          type: string
        modified:
          type: string
        title:
          type: array
          items:
            $ref: './langue-contenu.schema.json'
        uri:
          type: string
    Dataset:
      type: object
      properties:
        id:
          type: string
        uri:
          type: string
        catalogRecordCreated:
          type: string
        catalogRecordModified:
          type: string
        catalogRecordCreator:
          type: string
        catalogRecordContributor:
          type: string
        identifier:
          type: string
        issued:
          type: string
        modified:
          type: string
        disseminationStatus:
          type: string
        validationState:
          type: string
        version:
          type: string
        creator:
          type: array
          items:
            $ref: '#/components/schemas/IdLabel'
        publisher:
          $ref: '#/components/schemas/IdLabel'
        title:
          type: array
          items:
            $ref: './langue-contenu.schema.json'
        subtitle:
          type: array
          items:
            $ref: './langue-contenu.schema.json'
        abstract:
          type: array
          items:
            $ref: './langue-contenu.schema.json'
        description:
          type: array
          items:
            $ref: './langue-contenu.schema.json'
        scopeNote:
          type: array
          items:
            $ref: './langue-contenu.schema.json'
        wasGeneratedBy:
          type: array
          items:
            $ref: '#/components/schemas/IdLabel'
        type:
          type: array
          items:
            $ref: './langue-contenu.schema.json'
        archiveUnit:
          type: array
          items:
            $ref: '#/components/schemas/IdLabel'
        accessRights:
          type: array
          items:
            $ref: './langue-contenu.schema.json'
        confidentialityStatus:
          type: array
          items:
            $ref: './langue-contenu.schema.json'
        theme:
          type: array
          items:
            $ref: '#/components/schemas/Theme'
        landingPage:
          type: array
          items:
            $ref: '#/components/schemas/LocalisedUrl'
        processStep:
          $ref: '#/components/schemas/ProcessStep'
        accrualPeriodicity:
          $ref: 'langue-contenu.schema.json'
        temporal:
          $ref: '#/components/schemas/Temporal'
        temporalResolution:
          type: array
          items:
            $ref: 'langue-contenu.schema.json'
        spatial:
          $ref: '#/components/schemas/IdLabel'
        spatialResolution:
          type: array
          items:
            $ref: '#/components/schemas/IdLabel'
        spatialTemporal:
          type: string
        statisticalUnit:
          type: array
          items:
            $ref: '#/components/schemas/IdLabel'
        structure:
          $ref: '#/components/schemas/Structure'
        numObservations:
          type: integer
          format: int32
        numSeries:
          type: integer
          format: int32
        wasDerivedFrom:
          $ref: '#/components/schemas/WasDerivedFrom'
        relations:
          type: array
          items:
            type: string
        keyword:
          type: array
          items:
            $ref: './langue-contenu.schema.json'
    ListeDatasets:
      description: 'Liste des datasets'
      type: array
      xml:
        wrapped: true
        name: Definition
      items:
        $ref: '#/components/schemas/Dataset'